# .openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CWS Next App API",
    "version": "0.1.0",
    "description": "Automated API documentation for the CWS Next App. This document is generated from Zod schemas and route handler metadata.",
    "contact": {
      "name": "API Support",
      "email": "api@crossweavesourcing.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/api/mobile/v1/admin/products": {
      "post": {
        "tags": [
          "Mobile Admin"
        ],
        "summary": "Create a new product",
        "description": "Admin only. Creates a new product.",
        "operationId": "createProductMobile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoryId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  },
                  "shortDescription": {
                    "type": "string"
                  },
                  "overview": {
                    "type": "string"
                  },
                  "visible": {
                    "type": "boolean"
                  },
                  "manufacturing": {
                    "type": "string",
                    "description": "JSON stringified array"
                  },
                  "features": {
                    "type": "string",
                    "description": "JSON stringified array"
                  },
                  "specifications": {
                    "type": "string",
                    "description": "JSON stringified object"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": [
                  "name",
                  "slug",
                  "shortDescription",
                  "overview",
                  "visible",
                  "image"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "productId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires admin role",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cws_session",
        "description": "HMAC-signed session cookie issued by the web dashboard login flow. Obtained via POST /api/auth/login (Server Action) or Google OAuth callback."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ed25519-signed JWT issued by the mobile authentication endpoints. Obtained via POST /api/mobile/v1/auth/password or /api/mobile/v1/auth/google."
      },
      "pendingSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cws_2fa_pending",
        "description": "Short-lived HMAC-signed pending session cookie set during MFA challenge. Valid for 5 minutes. Used by WebAuthn and TOTP verification endpoints."
      }
    },
...
  }
}